feat: Add CameraController
